Subdivided the collapsable section from the character card
